+++ index.html
@@ -165,5 +165,6 @@
     </section>
   </template>
 
+  <script src="/js/main.js" type="module"></script>
 </body>
 </html>
