+++ package.json
@@ -3,6 +3,7 @@
   "version": "27.0.0",
   "private": true,
   "description": "Личный проект «Кекстаграм (простой)» от HTML Academy",
+  "type": "module",
   "repository": {
     "type": "git",
     "url": "git@github.com:htmlacademy-javascript/2233345-kekstagram-simple-27.git"
